New approach: Remove StructValue.__getattr__
